requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bulb-channels==2.4.0.1
celery==5.2.7
certifi==2019.9.11
cffi==1.15.1
channels==2.4.0
channels-redis==3.4.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap4==22.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.14.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
msgpack==1.0.4
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.1.0
pytz==2019.2
redis==4.3.4
redis-server==6.0.9
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
soupsieve==2.3.2.post1
sqlparse==0.3.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.25.6
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.4.1
whitenoise==4.1.4
wrapt==1.14.1
zope.interface==5.4.0
